refactor: extract bootstrap deep-link and verification intercept stages from inbound handler
